Crate rust_webvr

Modules

api

mock

utils

vr_display

vr_display_capabilities

vr_display_data

vr_event

vr_eye

vr_eye_parameters

vr_field_view

vr_frame_data

vr_framebuffer

vr_future_frame_data

vr_gamepad

vr_layer

vr_main_thread_heartbeat

vr_manager

vr_pose

vr_service

vr_stage_parameters

Structs

GlWindowVRService

MockVRInit

MockVRView

VRDisplayCapabilities
describes the capabilities of a VRDisplay. These are expected to be static per-device/per-user.

VRDisplayData

VREyeParameters
The VREyeParameters interface represents all the information required to correctly render a scene for a given eye.

VRFieldOfView
The VRFieldOfView interface represents a field of view, as given by 4 degrees describing the view from a center point.

VRFrameData
Represents all the information needed to render a single frame of a VR scene

VRFramebuffer
Information about a FBO provided by a VRDisplay.

VRFramebufferAttributes

VRFutureFrameData

VRGamepadButton

VRGamepadData

VRGamepadState

VRLayer
Data provided to a VRDisplay and presented in the HMD.

VRPose
The VRPose struct represents a sensor’s state at a given timestamp.

VRResolveFrameData

VRServiceManager

VRStageParameters
The VRStageParameters interface represents the values describing the stage/play area for displays that support room-scale experiences.

VRViewport

Enums

MockVRControlMsg

VRDisplayEvent

VRDisplayEventReason

VREvent

VREye
VREye

VRGamepadEvent

VRGamepadHand

Traits

VRDisplay
The VRDisplay traits forms the base of all VR device implementations

VRGamepad

VRMainThreadHeartbeat
Each VR service may have some code which needs to run on the main thread, for example window creation on MacOS is only supported on the main thread. Implementations of this trait will usually be neither Sync nor Send.

VRService

VRServiceCreator

Type Definitions

VRDisplayPtr

VRGamepadPtr