Trait rust_webvr::VRMainThreadHeartbeat

pub trait VRMainThreadHeartbeat {
    fn heartbeat(&mut self);

    fn heart_racing(&self) -> bool;
}

Each VR service may have some code which needs to run on the main thread, for example window creation on MacOS is only supported on the main thread. Implementations of this trait will usually be neither Sync nor Send.

Required methods

fn heartbeat(&mut self)

Run the heartbeat on the main thread.

fn heart_racing(&self) -> bool

Is the heartbeat expecting to be called every frame?

Loading content...